requirements.txt
flask
dotenv
gunicorn
psycopg2-binary
requests
flask-mail